+++ src/app/globals.css
@@ -67,5 +67,17 @@ h1{
     text-align: center;
 }
 
+.bg-color{
+  background-color: rgba(30, 30, 30, 0.75); /* dark grey with no purple tint */
+  backdrop-filter: blur(6px);
+  box-shadow: 0 8px 24px rgba(0, 0, 0, 0.4); /* solid black shadow */
+  border-radius: 0.75rem;     
+}
+
+
+.box-footer{
+  box-shadow: 6px 0px 0px hsl(0deg 0% 0% / 0.44);
+}
+
